[ENH] refactor forecasting metrics into individual private submodules (#8279)
This PR splits up the "mega-file" which contained all 22 forecasting
metric classes into individual private files.

There are no changes to public interfaces.
