+++ docs/source/api_reference/performance_metrics.rst
@@ -26,7 +26,35 @@ Forecasting
 Point forecasts - classes
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
+Average losses
+^^^^^^^^^^^^^^
+
+.. autosummary::
+    :toctree: auto_generated/
+    :template: class_with_call.rst
+
+    MeanAbsoluteError
+    MeanSquaredError
+    MedianAbsoluteError
+    MedianSquaredError
+
+Percentage errors
+^^^^^^^^^^^^^^^^^
+
 .. currentmodule:: sktime.performance_metrics.forecasting
+
+.. autosummary::
+    :toctree: auto_generated/
+    :template: class_with_call.rst
+
+    MeanAbsolutePercentageError
+    MedianAbsolutePercentageError
+    MeanSquaredPercentageError
+    MedianSquaredPercentageError
+    MeanSquaredErrorPercentage
+
+Scaled errors
+^^^^^^^^^^^^^
 
 .. autosummary::
     :toctree: auto_generated/
@@ -36,23 +64,40 @@ Point forecasts - classes
     MedianAbsoluteScaledError
     MeanSquaredScaledError
     MedianSquaredScaledError
-    MeanAbsoluteError
-    MeanSquaredError
-    MedianAbsoluteError
-    MedianSquaredError
-    GeometricMeanAbsoluteError
-    GeometricMeanSquaredError
-    MeanAbsolutePercentageError
-    MedianAbsolutePercentageError
-    MeanSquaredPercentageError
-    MedianSquaredPercentageError
+
+Relative errors
+^^^^^^^^^^^^^^^
+
+.. autosummary::
+    :toctree: auto_generated/
+    :template: class_with_call.rst
+
     MeanRelativeAbsoluteError
     MedianRelativeAbsoluteError
+    RelativeLoss
+
+Geometric errors
+^^^^^^^^^^^^^^^^
+
+.. autosummary::
+    :toctree: auto_generated/
+    :template: class_with_call.rst
+
+    GeometricMeanAbsoluteError
+    GeometricMeanSquaredError
     GeometricMeanRelativeAbsoluteError
     GeometricMeanRelativeSquaredError
+
+Under- and over-prediction errors
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autosummary::
+    :toctree: auto_generated/
+    :template: class_with_call.rst
+
     MeanAsymmetricError
     MeanLinexError
-    RelativeLoss
+
 
 Point forecasts - functions
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
