+++ sktime/performance_metrics/forecasting/probabilistic/_classes.py
@@ -8,7 +8,7 @@ from pandas.api.types import is_numeric_dtype
 from sklearn.utils import check_array, check_consistent_length
 
 from sktime.datatypes import check_is_scitype, convert, convert_to
-from sktime.performance_metrics.forecasting._classes import BaseForecastingErrorMetric
+from sktime.performance_metrics.forecasting._base import BaseForecastingErrorMetric
 from sktime.performance_metrics.forecasting._coerce import _coerce_to_scalar
 
 # TODO: Rework tests now
